Finish

+++ src/deposit.h
@@ -2,6 +2,6 @@
 #define DEPOSIT_H
 
 int check(int day, int summa);
-float calc(int day, int summa);
+double calc(int day, int summa);
 
 #endif
